requirements.txt
aenum==3.1.16
afip.py==1.2.0
altgraph==0.17.4
bcrypt==5.0.0
blinker==1.9.0
branca==0.8.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy>=1.2.0
cryptography==46.0.2
cycler==0.12.1
dbf==0.99.11
et_xmlfile==2.0.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
folium==0.20.0
fonttools==4.59.2
fpdf==1.7.2
future==1.0.0
geographiclib==2.1
geopandas==1.1.1
geopy==2.4.1
gunicorn==23.0.0
httplib2==0.31.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.7
narwhals==2.5.0
numpy==1.26.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
patsy==1.0.1
pefile==2023.2.7
pillow==11.3.0
plotly==6.3.0
psycopg2-binary==2.9.10
pycparser==2.23
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
PyJWT==2.10.1
pyogrio==0.11.1
pyparsing==3.2.3
pyproj==3.7.2
PySimpleSOAP==1.8.22
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
qrcode==8.2
requests==2.32.5
scipy==1.15.0
seaborn==0.13.2
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
statsmodels==0.14.5
tabulate==0.9.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
xlsxwriter==3.2.8
xyzservices==2025.4.0
# Para cargar variables de entorno desde un archivo .env en desarrollo local
python-dotenv>=1.0.0
Flask-Mail==0.10.0
APScheduler>=3.10.0
mkdocs-material>=9.0.0
mkdocs-awesome-pages-plugin>=2.9.0
